+++ python/lsst/ts/mtaos/mtaos.py
@@ -70,7 +70,7 @@ class MTAOS(salobj.ConfigurableCsc):
     version = __version__
 
     DEFAULT_TIMEOUT = 10.0
-    LONG_TIMEOUT = 60.0
+    LONG_TIMEOUT = 90.0
     LOG_FILE_NAME = "MTAOS.log"
     MAX_TIME_SAMPLE = 100
     CMD_TIMEOUT = 60.0
@@ -239,6 +239,10 @@ class MTAOS(salobj.ConfigurableCsc):
         # loop.
         self.last_run_ofc_configuration = ""
         self.image_rotator = dict()
+        self.current_elevation_position = None
+        self.current_rotator_position = None
+
+        self.previous_dofs = None
 
         self.log.info("MTAOS CSC is ready.")
 
@@ -348,7 +352,9 @@ class MTAOS(salobj.ConfigurableCsc):
             ),
         )
 
-        if dof_state0 is not None:
+        if self.previous_dofs is not None:
+            self.model.ofc.aggregated_state = self.previous_dofs
+        elif dof_state0 is not None:
             self.model.ofc_data.dof_state0 = dof_state0
 
         if hasattr(config, "wep_config"):
@@ -381,6 +387,10 @@ class MTAOS(salobj.ConfigurableCsc):
         self.stress_scale_factor = config.stress_scale_factor
         self.m1m3_stress_limit = config.m1m3_stress_limit
         self.m2_stress_limit = config.m2_stress_limit
+
+        # Set elevation and rotation angle limits
+        self.elevation_angle_limit = config.elevation_delta_limit_max
+        self.rotation_angle_limit = config.rotation_delta_limit
 
         self.log.debug("MTAOS configuration completed.")
 
@@ -437,6 +447,92 @@ class MTAOS(salobj.ConfigurableCsc):
                 pass
 
         await self.evt_closedLoopState.set_write(state=ClosedLoopState.IDLE)
+
+    async def begin_start(self, data):
+        await self.cmd_start.ack_in_progress(
+            data,
+            timeout=self.CMD_TIMEOUT,
+            result="MTAOS CSC started.",
+        )
+        await super().begin_start(data)
+
+    async def begin_enable(self, data):
+        await self.cmd_enable.ack_in_progress(
+            data,
+            timeout=self.CMD_TIMEOUT,
+            result="Enabling MTAOS CSC.",
+        )
+        await super().begin_enable(data)
+
+    async def handle_summary_state(self):
+        """Handle summary state changes.
+        Here we store the previous state of the DOFs when
+        going to fault or disabled.
+        """
+        if self.summary_state in {salobj.State.FAULT, salobj.State.DISABLED}:
+            self.log.info("Storing previous state.")
+            self.previous_dofs = self.model.ofc.controller.aggregated_state
+
+        elif self.summary_state is salobj.State.ENABLED:
+            self.log.info("Restoring previous state.")
+            try:
+                if (
+                    await self.check_components_alive()
+                    and await self.check_components_enabled()
+                ):
+                    await self._execute_issue_correction()
+                else:
+                    self.log.warning(
+                        "One or more CSCs are not alive or enabled, skip issuing correction."
+                    )
+            except Exception:
+                self.log.exception(
+                    "MTAOS unable to apply initial corrections. Ignoring."
+                )
+
+    async def check_components_enabled(self):
+        """Checks if all components are ENABLED.
+
+        Raises
+        ------
+        RunTimeError:
+            If either component is not ENABLED
+        """
+        return all(
+            await asyncio.gather(
+                *[self._check_enabled(remote) for remote in self.remotes]
+            )
+        )
+
+    async def _check_enabled(self, remote) -> bool:
+        """Check if a specific remote component is enabled."""
+        summary_state = await self.remote[remote].evt_summaryState.aget()
+        return salobj.State(summary_state.summaryState) == salobj.State.ENABLED
+
+    async def check_components_alive(self) -> bool:
+        """Check if all AOS components are alive, e.g. publishing heartbeats.
+
+        Returns
+        -------
+        `bool`
+            True if all components are alive, False otherwise
+        """
+        return all(
+            await asyncio.gather(
+                *[self._check_liveliness(remote) for remote in self.remotes]
+            )
+        )
+
+    async def _check_liveliness(self, remote) -> bool:
+        """Check if a specific remote component is alive."""
+        try:
+            await self.remote[remote].evt_heartbeat.next(
+                flush=True, timeout=self.DEFAULT_TIMEOUT
+            )
+        except asyncio.TimeoutError:
+            return False
+        else:
+            return True
 
     async def do_resetCorrection(self, data):
         """Command to reset the current wavefront error calculations.
@@ -946,6 +1042,11 @@ class MTAOS(salobj.ConfigurableCsc):
     def do_startClosedLoop(self, data):
         """Start the closed loop operation."""
         self.assert_enabled()
+
+        # Set ofc configuration to be used in closed loop.
+        if data.config:
+            self.last_run_ofc_configuration = data.config
+
         if self.closed_loop_task.done():
             self.closed_loop_task = asyncio.create_task(self.run_closed_loop())
         else:
@@ -961,6 +1062,8 @@ class MTAOS(salobj.ConfigurableCsc):
         self.log.info(f"Starting closed loop for {oods_name}.")
 
         processed_images = collections.deque(maxlen=100)
+        self.current_rotator_position = None
+        self.current_elevation_position = None
 
         async with salobj.Remote(
             self.domain,
@@ -980,13 +1083,19 @@ class MTAOS(salobj.ConfigurableCsc):
             "MTRotator",
             readonly=True,
             include=["summaryState", "rotation"],
-        ) as mtrotator:
+        ) as mtrotator, salobj.Remote(
+            self.domain,
+            "MTMount",
+            readonly=True,
+            include=["summaryState", "elevation"],
+        ) as mtmount:
 
             self.log.info("Closed loop task ready.")
 
             camera.evt_startIntegration.callback = self.follow_start_integration
             camera.evt_endOfImageTelemetry.callback = self.follow_end_integration
             mtrotator.tel_rotation.callback = self.follow_rotator_position
+            mtmount.tel_elevation.callback = self.follow_elevation_position
 
             oods.evt_imageInOODS.flush()
             while self.summary_state == salobj.State.ENABLED:
@@ -1006,6 +1115,56 @@ class MTAOS(salobj.ConfigurableCsc):
                         self.log.info(f"Visit {visit_id} already processed, skipping.")
                         continue
 
+                    filter_label, elevation = await self.model.get_image_info(
+                        visit_id,
+                        self.camera_name,
+                    )
+
+                    rotation_angle = float(
+                        np.mean(np.array(self.image_rotator[image_in_oods.obsid]))
+                    )
+
+                    # Since the current rotator and elevation positions
+                    # are being updated in the background, we need to
+                    # create a local copy and assume it's static
+                    # in the following if statement.
+                    current_rotator_position = self.current_rotator_position
+                    current_elevation_position = self.current_elevation_position
+
+                    if (
+                        current_elevation_position is None
+                        or current_rotator_position is None
+                        or (
+                            (
+                                np.abs(rotation_angle - current_rotator_position)
+                                > self.rotation_angle_limit
+                            )
+                            or (
+                                np.abs(elevation - current_elevation_position)
+                                > self.elevation_angle_limit
+                            )
+                        )
+                    ):
+                        report_elevation = (
+                            f"{current_elevation_position:.1f}"
+                            if current_elevation_position is not None
+                            else "not set"
+                        )
+                        report_rotator = (
+                            f"{current_rotator_position:.1f}"
+                            if current_rotator_position is not None
+                            else "not set"
+                        )
+                        self.log.info(
+                            f"Current rotator position: {report_rotator:.1f}, "
+                            f"current elevation: {report_elevation:.1f}. "
+                            f"Image to process was taken at {rotation_angle=} and {elevation=}. "
+                            f"Difference above threshold rot_limit={self.rotation_angle_limit} "
+                            f"el_limit={self.elevation_angle_limit}. "
+                            "Skipping."
+                        )
+                        continue
+
                     processed_images.append(visit_id)
 
                     self.log.info(
@@ -1022,15 +1181,6 @@ class MTAOS(salobj.ConfigurableCsc):
                         extra_id=None,
                         use_ocps=self.use_ocps,
                         config=self.wep_config,
-                    )
-
-                    filter_label, elevation = await self.model.get_image_info(
-                        visit_id,
-                        self.camera_name,
-                    )
-
-                    rotation_angle = float(
-                        np.mean(np.array(self.image_rotator[image_in_oods.obsid]))
                     )
 
                     gain = await self.model.get_correction_gain(
@@ -1727,8 +1877,12 @@ class MTAOS(salobj.ConfigurableCsc):
                 self.image_rotator.pop(item)
 
     async def follow_rotator_position(self, data):
+        self.current_rotator_position = data.actualPosition
         if self.current_image is not None:
             self.image_rotator[self.current_image].append(data.actualPosition)
+
+    async def follow_elevation_position(self, data):
+        self.current_elevation_position = data.actualPosition
 
     def get_subsystems_versions(self) -> str:
         """Get subsystems versions string.